Separate config loading steps and applying presets into diff functions
